Move FilterGambits to a Query namespace
